Change the folder structure for constant file
