+++ src/constant/columns/expenseColumns.js
@@ -1,4 +1,4 @@
-import { formatAmount, formatDate, formatTransactionType } from "./helper";
+import { formatAmount, formatDate, formatTransactionType } from "../helper";
 
 const expenseColumns = [
   { title: "Name", key: "name" },
