+++ src/constant/columns/transactionColumns.js
@@ -1,4 +1,4 @@
-import { formatDate } from "./helper";
+import { formatDate } from "../helper";
 
 const transactionColumns = [
   { title: "Image", key: "url" },
