+++ src/constant/dropdowns/reportDropdown.js
@@ -1,0 +1,12 @@
+const reportDropdown = [
+  {
+    id: 1,
+    name: "category",
+    placeholder: "All",
+    label: "Select Category",
+    className: "w-31",
+    options: [],
+  },
+];
+
+export default reportDropdown;
