+++ src/constant/inputs/categoryInputs.js
@@ -1,0 +1,11 @@
+const categoryInputs = [
+  {
+    id: 1,
+    name: "name",
+    type: "text",
+    placeholder: "Bills.",
+    label: "name",
+  },
+];
+
+export default categoryInputs;
