+++ src/constant/inputs/expenseInput.js
@@ -28,26 +28,4 @@ const expenseInputs = [
   },
 ];
 
-const expenseSelects = [
-  {
-    id: 1,
-    name: "type",
-    placeholder: "Select Transaction",
-    label: "Transaction Type",
-    className: "order-1",
-    options: [
-      { value: "expense", name: "Expense" },
-      { value: "income", name: "Income" },
-    ],
-  },
-  {
-    id: 2,
-    name: "category",
-    placeholder: "Select Category",
-    label: "Select Category",
-    className: "order-2",
-    options: [{ value: "", name: "Select Category" }],
-  },
-];
-
-export { expenseInputs, expenseSelects };
+export default expenseInputs;
