+++ src/screens/account-setting/index.jsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
 import { Button, Card, CardBody, CardTitle, Form, Spinner } from "reactstrap";
 import { toast } from "react-toastify";
-import Input from "../../components/Input";
 import { useDispatch, useSelector } from "react-redux";
-import { editUserReducer } from "../../feature/auth/userSlice";
-import { storage } from "../../config/firebaseConfig";
-import accountSettingInputs from "../../config/constant/accountSettingInputs";
+import Input from "../../components/Input";
 import useUser from "../../hooks/useUser";
+import { editUserReducer } from "../../feature/auth/userSlice";
+import accountSettingInputs from "../../constant/inputs/accountSettingInputs";
+import { storage } from "../../config/firebaseConfig";
 import {
   updateUser,
   updateUserWithImage,
